refactor(daemon): share command and teardown budgets
