+++ src/daemon/server/daemon-runtime.ts
@@ -1,3 +1,4 @@
+import { resolveDaemonSessionTeardownTimeoutMs } from '../session-teardown-budget.ts';
 import crypto from 'node:crypto';
 import { asAppError, AppError } from '@agent-device/kernel/errors';
 import { resolveSessionRequestLogPath, SessionStore } from '../session-store.ts';
@@ -74,7 +75,7 @@ import {
   platformResourceCleanup,
   resetAndroidSnapshotHelperRuntime,
 } from '../../platform-runtime-resource-cleanup.ts';
-import { isWebSession, openWebSessionNames } from '../web-session-names.ts';
+import { openWebSessionNames } from '../web-session-names.ts';
 import {
   recoverAppLogResourcesAfterDaemonLock,
   type AppLogRecoveryDiagnostic,
@@ -84,12 +85,6 @@ import { createAppLogAdmissionLedger } from '../app-log-admission-ledger.ts';
 import { createAudioProbeAdmissionLedger } from '../audio-probe-admission-ledger.ts';
 import { createScreenRecordingAdmissionLedger } from '../screen-recording-admission-ledger.ts';
 
-const DAEMON_SESSION_TEARDOWN_TIMEOUT_MS = 5_000;
-export const SCREEN_RECORDING_SESSION_TEARDOWN_BUDGET_MS = 11_000;
-// Mirrors AGENT_BROWSER_TIMEOUT_MS in @agent-device/platform-web: the ceiling that
-// module already places on one `agent-browser` CLI call, so the race below never gives up on the
-// web-close step while the close it started is still running within its own enforced limit.
-export const WEB_BROWSER_SESSION_TEARDOWN_BUDGET_MS = 30_000;
 const DAEMON_SESSION_LEASE_RELEASE_TIMEOUT_MS = 1_000;
 const DAEMON_PNG_WORKER_TERMINATE_TIMEOUT_MS = 1_000;
 const DAEMON_PROVIDER_RELEASE_DRAIN_TIMEOUT_MS = 2_000;
@@ -97,20 +92,6 @@ const DAEMON_PROVIDER_RELEASE_DRAIN_TIMEOUT_MS = 2_000;
 type WritableOutput = {
   write: (chunk: string) => unknown;
 };
-
-/**
- * Per-session teardown budget for daemon shutdown. The base budget covers ordinary resources; an
- * active durable recording or an open web session gets an additional owner-finalization budget so
- * the daemon cannot exit while its runtime handle is still producing the terminal media artifact,
- * or while agent-browser is still closing its Chrome fleet. The base portion remains available to
- * cleanup steps that follow recording finalization or the web close.
- */
-export function resolveDaemonSessionTeardownTimeoutMs(session: SessionState): number {
-  let timeoutMs = DAEMON_SESSION_TEARDOWN_TIMEOUT_MS;
-  if (session.screenRecording) timeoutMs += SCREEN_RECORDING_SESSION_TEARDOWN_BUDGET_MS;
-  if (isWebSession(session)) timeoutMs += WEB_BROWSER_SESSION_TEARDOWN_BUDGET_MS;
-  return timeoutMs;
-}
 
 async function settleDaemonTeardownStep(params: {
   session: SessionState;
